// src-old/types/core/UnitInterfaces.ts
export interface IFixedAllocation {
  readonly id: EntityId;
  readonly name: string; // "XL Engine", "Endo Steel"
  readonly location: string;
  readonly slotIndex: number;
  readonly slots: number; // Number of slots occupied
  readonly type: 'engine' | 'gyro' | 'cockpit' | 'structure' | 'armor' | 'heatsink' | 'other';
}